Feat: Move SimpleGUI, SettingsBase to Helper
