+++ Helper/Abstraction/SettingsBase.cs
@@ -5,14 +5,13 @@ using System.Text;
 using System.Threading.Tasks;
 using SimpleJSON;
 
-namespace NoStopMod.Abstraction
+namespace NoStopMod.Helper.Abstraction
 {
     interface SettingsBase
     {
-
         void Load(ref JSONNode json);
 
         void Save(ref JSONNode json);
-        
+
     }
 }
