+++ Helper/SimpleGUI.cs
@@ -5,7 +5,7 @@ using System.Text;
 using System.Threading.Tasks;
 using UnityEngine;
 
-namespace NoStopMod.CustomLayout
+namespace NoStopMod.Helper
 {
     class SimpleGUI
     {
